Move game output messages to their own file
