+++ game/src/lib.rs
@@ -5,3 +5,4 @@ pub mod items;
 pub mod messages;
 pub mod threats;
 pub mod tributes;
+pub mod output;
